Do the project iteration on Gerrit startup in background
On Gerrit startup the plugin iterates over all projects and schedules
for each project a background job to scan for project-specific
download commands. Although the actual scanning is done in background
the iteration over all projects and the scheduling of the background
jobs may take a long time. Improve the Gerrit startup time by moving
the iteration over all projects into a background job that then does
the scanning for project-specific download commands on all projects.

Change-Id: I38e6005568f990ec816f4af43f40cb77091a2483
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>